fix(email-notifications): use code string literals for default content
